requirements.txt
# Geo6D-Lite Pose Estimation


# Core deep learning libraries
torch>=2.1.0
torchvision>=0.16.0

# Restrict numpy to avoid conflicts with TF/Streamlit/Numba
numpy>=1.26.0,<2.0.0
scipy>=1.11.0

# Image and visualization
opencv-python>=4.10.0
matplotlib>=3.8.0
Pillow>=10.0.0

# Progress bars and utilities
tqdm>=4.66.0

# Data loading / file handling
pyyaml>=6.0.1

# API / deployment
flask>=3.0.0
requests>=2.31.0

# Optional: plotting and Colab visualization
seaborn>=0.13.0

# (Colab-safe version of OpenCV)
opencv-python-headless>=4.10.0.84

Notes

trimesh>=3.22.0
psutil>=5.9.0
flask>=3.0.0